+++ jitsi-utils-kotlin/src/main/kotlin/org/jitsi/utils/configk2/dsl/ConfigPropertyDsl.kt
@@ -21,7 +21,6 @@ import org.jitsi.utils.configk2.ConfigPropertyAttributes
 import org.jitsi.utils.configk2.ConfigPropertyAttributesBuilder
 import org.jitsi.utils.configk2.ConfigRetriever
 import org.jitsi.utils.configk2.ConfigSource
-import java.time.Duration
 import kotlin.reflect.KClass
 
 class ConfigPropertyBuilder<T : Any>(
@@ -82,36 +81,4 @@ class ConfigPropertyBuilder<T : Any>(
 inline fun <reified T : Any> property(block: ConfigPropertyBuilder<T>.() -> Unit): ConfigProperty<T> {
     val x = ConfigPropertyBuilder(T::class).also(block)
     return x.build()
-}
-
-class DummyConfig : ConfigSource {
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : Any> getterFor(valueType: KClass<T>): (ConfigSource, String) -> T {
-        return when(valueType) {
-            Int::class -> ({ config, path -> (config as DummyConfig).getInt(path) as T })
-            Duration::class -> ({ config, path -> (config as DummyConfig).getDuration(path) as T })
-            else -> TODO("no getter for $valueType")
-        }
-    }
-
-    fun getInt(path: String): Int = 42
-    fun getDuration(path: String): Duration = Duration.ofSeconds(10)
-}
-
-fun main() {
-    val x = property<Long> {
-        name("name")
-        readOnce()
-        fromConfig(DummyConfig())
-        retrievedAs<Duration>() convertedBy { it.toMillis() }
-    }
-
-    val y = property<Int> {
-        name("name")
-        readOnce()
-        fromConfig(DummyConfig())
-    }
-
-    println("x = ${x.value}")
-    println("y = ${y.value}")
 }
